third_party/blink/web_tests/platform/generic/fast/js/post-inc-assign-overwrites-expected.txt

Test for regression against

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS postIncDotAssignToBase() is true
PASS postIncBracketAssignToBase() is true
PASS postIncBracketAssignToSubscript() is true
PASS successfullyParsed is true

TEST COMPLETE
